outputs/generated_repo/car-webssitee-showing-all-toyota-cars-sold-in-pakistan/main.py
```python
# main.py

# Import required libraries
import os
import sys
import requests
from flask import Flask, render_template, request
from flask_cors import CORS
from github import Github
from datetime import datetime

# Create a Flask application instance
app = Flask(__name__)
CORS(app)

# Define a dictionary to store Toyota car models sold in Pakistan
toyota_cars = {
    "Corolla": {"price": "3,500,000 - 4,500,000 PKR", "description": "A compact sedan"},
    "Camry": {"price": "5,000,000 - 6,500,000 PKR", "description": "A mid-size sedan"},
    "Hilux": {"price": "3,000,000 - 4,000,000 PKR", "description": "A pickup truck"},
    "Fortuner": {"price": "5,500,000 - 7,000,000 PKR", "description": "A mid-size SUV"},
    "Prius": {"price": "4,000,000 - 5,000,000 PKR", "description": "A hybrid compact car"},
}

# Define a function to generate project specification
def generate_project_specification():
    """
    Generate project specification for the car website.
    """
    specification = {
        "Project Name": "Car Website",
        "Description": "A website showing all Toyota cars sold in Pakistan",
        "Features": ["Display Toyota car models", "Display car prices", "Display car descriptions"],
        "Target Users": ["Developers", "Maintainers", "Contributors"],
    }
    return specification

# Define a function to scaffold a repository
def scaffold_repository():
    """
    Scaffold a repository for the car website.
    """
    # Create a new repository on GitHub
    g = Github("your-github-token")
    user = g.get_user()
    repo = user.create_repo("car-website", description="A website showing all Toyota cars sold in Pakistan")

    # Create a new branch
    repo.create_git_ref(ref="refs/heads/feature/new-feature", sha=repo.get_branch("main").commit.sha)

    # Create a new file
    repo.create_file("README.md", "Initial commit", "# Car Website\nA website showing all Toyota cars sold in Pakistan", branch="feature/new-feature")

# Define a function to draft documentation
def draft_documentation():
    """
    Draft documentation for the car website.
    """
    documentation = {
        "Introduction": "This is a website showing all Toyota cars sold in Pakistan.",
        "Features": ["Display Toyota car models", "Display car prices", "Display car descriptions"],
        "Usage": ["Open the website in a web browser", "Select a car model to view its details"],
    }
    return documentation

# Define a function to set up GitHub workflow
def setup_github_workflow():
    """
    Set up GitHub workflow for the car website.
    """
    # Create a new workflow file
    workflow_file = """
name: Build and deploy
on:
  push:
    branches:
      - main
jobs:
  build-and-deploy:
    runs-on: ubuntu-latest
    steps:
      - name: Checkout code
        uses: actions/checkout@v2
      - name: Install dependencies
        run: |
          pip install -r requirements.txt
      - name: Build and deploy
        run: |
          python main.py
    """
    with open(".github/workflows/main.yml", "w") as f:
        f.write(workflow_file)

# Define a route for the home page
@app.route("/")
def home():
    """
    Display the home page.
    """
    return render_template("index.html", toyota_cars=toyota_cars)

# Define a route for the car details page
@app.route("/car/<car_model>")
def car_details(car_model):
    """
    Display the car details page.
    """
    car = toyota_cars.get(car_model)
    if car:
        return render_template("car_details.html", car_model=car_model, car=car)
    else:
        return "Car not found", 404

# Define a route for the project specification page
@app.route("/project-specification")
def project_specification():
    """
    Display the project specification page.
    """
    specification = generate_project_specification()
    return render_template("project_specification.html", specification=specification)

# Define a route for the repository scaffolding page
@app.route("/scaffold-repository")
def scaffold_repository_page():
    """
    Scaffold a repository for the car website.
    """
    scaffold_repository()
    return "Repository scaffolded successfully"

# Define a route for the documentation drafting page
@app.route("/draft-documentation")
def draft_documentation_page():
    """
    Draft documentation for the car website.
    """
    documentation = draft_documentation()
    return render_template